a new subpackage, which is lib.indexing, is added and lib.combo.arange.py is moved there; fftflow_.py updated to revise for this movement; spline.py is slightly modified for making sure that knots are squeezed and unsqueezed.
Toward Schwinger model. BTW, util/measure/su3_eigenphase.py is planned to be moved to normflow_util.

One important change: The beta of Softmax in nn/scalar/modules.py and nn/scalar/coupling_.py are set log(2) so that Softmax returns 1 for zero input. Moreover, it is now possible to use 'set_param2zero' method of ConvAct to set the parameters to zero.